feat: restructure and introduce question processing, validation logic, tag parsing and rendering mechanisms
