+++ backend/src/modules/quizzes/question-processing/renderers/index.ts
@@ -1,0 +1,3 @@
+export * from './interfaces/RenderViews';
+export * from './BaseQuestionRenderer';
+export * from './SOLQuestionRenderer';
