+++ backend/src/modules/quizzes/question-processing/tag-parser/index.ts
@@ -1,0 +1,5 @@
+export * from './TagParser';
+export * from './tags/Tag';
+export * from './tags/NumExprTag';
+export * from './tags/NumExprTexTag';
+export * from './tags/QParamTag';
