+++ backend/src/modules/quizzes/question-processing/validators/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseQuestionValidator';
+export * from './SOLQuestionValidator';
