--- backend/src/modules/quizzes/rendering/QuestionProcessor.ts
@@ -1,161 +1,0 @@
-import {BaseQuestion, SOLQuestion} from 'modules/quizzes/classes/transformers';
-import {TagParserEngine} from './TagParserEngine';
-import {QParamProcessor} from './processors/QParamProcessor';
-import {NumExprProcessor} from './processors/NumExprProcessor';
-import {NumExprTexProcessor} from './processors/NumExprTexProcessor';
-import {ILotItem} from 'shared/interfaces/quiz';
-import {BaseQuestionValidator} from '../rules/questions/BaseQuestionValidator';
-import {SelectOneInLotValidator} from '../rules/questions';
-import {triggerAsyncId} from 'async_hooks';
-import {ParameterMap} from './processors/ITagProcessor';
-import {generateRandomParameterMap} from '../utils/functions/generateRandomParameterMap';
-
-interface IQuestionRenderView extends BaseQuestion {
-  parameterMap?: ParameterMap;
-}
-
-interface SOLQuestionRenderView extends IQuestionRenderView {
-  lotItems: ILotItem[];
-}
-
-class BaseQuestionRenderer {
-  question: BaseQuestion;
-  tagParser: TagParserEngine;
-
-  constructor(question: BaseQuestion, tagParser: TagParserEngine) {
-    this.question = question;
-    this.tagParser = tagParser;
-  }
-
-  render(parameterMap: ParameterMap): BaseQuestion | IQuestionRenderView {
-    if (!this.question.isParameterized || !this.question.parameters?.length) {
-      return this.question;
-    }
-
-    const renderedQuestion: BaseQuestion = JSON.parse(
-      JSON.stringify(this.question),
-    );
-
-    // Apply a function for all these fields in question text, hint, correctItem.text, correctItem.hint
-    renderedQuestion.text = this.tagParser.processText(
-      this.question.text,
-      parameterMap,
-    );
-    if (this.question.hint) {
-      renderedQuestion.hint = this.tagParser.processText(
-        this.question.hint,
-        parameterMap,
-      );
-    }
-    return renderedQuestion;
-  }
-}
-
-class SOLQuestionRenderer extends BaseQuestionRenderer {
-  declare question: SOLQuestion;
-  declare tagParser: TagParserEngine;
-
-  constructor(question: SOLQuestion, tagParser: TagParserEngine) {
-    super(question, tagParser);
-  }
-  render(parameterMap: ParameterMap): SOLQuestionRenderView {
-    const renderedQuestion: SOLQuestion = super.render(
-      parameterMap,
-    ) as SOLQuestion;
-
-    const lotItems: ILotItem[] = [
-      renderedQuestion.correctLotItem,
-      ...renderedQuestion.incorrectLotItems,
-    ];
-    const processedLotItems = lotItems.map(item => ({
-      ...item,
-      text: this.tagParser.processText(item.text, parameterMap),
-      explaination: this.tagParser.processText(item.explaination, parameterMap),
-    }));
-
-    //Shuffle the lot
-    const shuffledLotItems = processedLotItems.sort(() => Math.random() - 0.5);
-
-    const renderedQuestionWithLotItems: SOLQuestionRenderView = {
-      _id: renderedQuestion._id,
-      type: renderedQuestion.type,
-      isParameterized: renderedQuestion.isParameterized,
-      text: renderedQuestion.text,
-      hint: renderedQuestion.hint,
-      points: renderedQuestion.points,
-      timeLimitSeconds: renderedQuestion.timeLimitSeconds,
-      lotItems: shuffledLotItems,
-      parameterMap: parameterMap,
-    };
-
-    return renderedQuestionWithLotItems;
-  }
-}
-
-class QuestionProcessor {
-  private tagParser: TagParserEngine;
-  private question: BaseQuestion;
-  private validator: BaseQuestionValidator;
-  private renderer: BaseQuestionRenderer;
-
-  private createValidator(): BaseQuestionValidator {
-    switch (this.question.type) {
-      case 'SELECT_ONE_IN_LOT':
-        return new SelectOneInLotValidator(
-          this.question as SOLQuestion,
-          this.tagParser,
-        );
-      // Add more cases for other question types as needed
-      default:
-        throw new Error(
-          `No validator found for question type: ${this.question.type}`,
-        );
-    }
-  }
-
-  private createRenderer(): BaseQuestionRenderer {
-    switch (this.question.type) {
-      case 'SELECT_ONE_IN_LOT':
-        return new SOLQuestionRenderer(
-          this.question as SOLQuestion,
-          this.tagParser,
-        );
-      // Add more cases for other question types as needed
-      default:
-        throw new Error(
-          `No renderer found for question type: ${this.question.type}`,
-        );
-    }
-  }
-
-  constructor(question: BaseQuestion) {
-    this.tagParser = new TagParserEngine({
-      QParam: new QParamProcessor(),
-      NumExpr: new NumExprProcessor(),
-      NumExprTex: new NumExprTexProcessor(),
-    });
-    this.question = question;
-    this.validator = this.createValidator();
-    this.renderer = this.createRenderer();
-  }
-
-  validate(): void {
-    this.validator.validate();
-  }
-
-  render(parameterMap?: ParameterMap): BaseQuestion | IQuestionRenderView {
-    if (parameterMap) {
-      return this.renderer.render(parameterMap);
-    }
-
-    // Generates a map of parameter names to their values
-    // This map is used to replace the parameter tags in the question text
-    // The values are randomly chosen
-    const randomParameterMap = generateRandomParameterMap(
-      this.question.parameters,
-    );
-    return this.renderer.render(randomParameterMap);
-  }
-}
-
-export {QuestionProcessor};
