--- backend/src/modules/quizzes/rules/questions/index.ts
@@ -1,2 +1,0 @@
-export * from './ParameterizedQuestionRules';
-export * from './SelectOneInLotValidator';
